split sidebar in leafs for easier optimizations for the react compiler
