+++ src/components/sidebar/app-sidebar.tsx
@@ -1,133 +1,28 @@
-"use client";
-
-import * as React from "react";
+import type { ComponentProps } from "react";
 import Link from "next/link";
 import Image from "next/image";
-import { usePathname } from "next/navigation";
-import { usePreloadedAuthQuery } from "@convex-dev/better-auth/nextjs/client";
-import { Preloaded } from "convex/react";
+import type { Preloaded } from "convex/react";
 import { api } from "@convex/_generated/api";
-import { usePostHogIdentify } from "@/hooks/use-posthog-identify";
-import {
-  MessageSquareIcon,
-  FolderIcon,
-  WrenchIcon,
-  PackageIcon,
-  BarChart2Icon,
-} from "lucide-react";
 import {
   Sidebar,
-  SidebarContent,
   SidebarFooter,
-  SidebarGroup,
-  SidebarGroupContent,
   SidebarHeader,
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import { NavUser } from "@/components/sidebar/nav-user";
-import { Separator } from "@/components/ui/separator";
-
-type Role = "admin" | "maker" | "client";
-
-type NavItem = {
-  title: string;
-  url: string;
-  icon: React.ReactNode;
-  iconBackground: string;
-  iconColor: string;
-  roles: Role[];
-  group: "main" | "manage" | "reports";
-};
-
-const allNavItems: NavItem[] = [
-  // ── Available to all ────────────────────────────────────────────────────
-  {
-    title: "Messages",
-    url: "/dashboard/chat",
-    icon: <MessageSquareIcon />,
-    iconBackground: "var(--fab-magenta-light)",
-    iconColor: "var(--fab-magenta)",
-    roles: ["admin", "maker", "client"],
-    group: "main",
-  },
-  {
-    title: "Projects",
-    url: "/dashboard/projects",
-    icon: <FolderIcon />,
-    iconBackground: "var(--fab-bg-main)",
-    iconColor: "var(--chart-4)",
-    roles: ["admin", "maker", "client"],
-    group: "main",
-  },
-  // ── Admin / Maker ────────────────────────────────────────────────────────
-  {
-    title: "Services",
-    url: "/dashboard/services",
-    icon: <WrenchIcon />,
-    iconBackground: "var(--fab-timeline-complete-soft)",
-    iconColor: "var(--fab-teal)",
-    roles: ["admin", "maker"],
-    group: "manage",
-  },
-  {
-    title: "Inventory",
-    url: "/dashboard/inventory",
-    icon: <PackageIcon />,
-    iconBackground: "var(--fab-amber-light)",
-    iconColor: "var(--fab-amber)",
-    roles: ["admin", "maker"],
-    group: "manage",
-  },
-  // ── Reports ──────────────────────────────────────────────────────────────
-  {
-    title: "Reports",
-    url: "#",
-    icon: <BarChart2Icon />,
-    iconBackground: "var(--fab-bg-deep)",
-    iconColor: "var(--fab-text-muted)",
-    roles: ["admin", "maker"],
-    group: "reports",
-  },
-];
-
-function filterNavItems(items: NavItem[], role: Role): NavItem[] {
-  return items.filter((item) => item.roles.includes(role));
-}
-
-function groupNavItems(items: NavItem[]): { key: string; items: NavItem[] }[] {
-  const ORDER = ["main", "manage", "reports"] as const;
-  const map: Record<string, NavItem[]> = {};
-  for (const item of items) {
-    (map[item.group] ??= []).push(item);
-  }
-  return ORDER.filter((k) => map[k]?.length > 0).map((k) => ({
-    key: k,
-    items: map[k],
-  }));
-}
+import {
+  SidebarNavigation,
+  SidebarUserFooter,
+} from "@/components/sidebar/app-sidebar-content";
+import { ProfileProvider } from "@/components/sidebar/profile-context";
 
 export function AppSidebar({
   preloadedProfile,
   ...props
-}: React.ComponentProps<typeof Sidebar> & {
+}: ComponentProps<typeof Sidebar> & {
   preloadedProfile: Preloaded<typeof api.users.getUserProfile>;
 }) {
-  const profile = usePreloadedAuthQuery(preloadedProfile);
-  const pathname = usePathname();
-  usePostHogIdentify(profile ?? null);
-
-  const role: Role = profile?.role ?? "client";
-  const navItems = filterNavItems(allNavItems, role);
-  const groups = groupNavItems(navItems);
-
-  const user = {
-    name: profile?.name ?? "",
-    email: profile?.email ?? "",
-    avatar: profile?.profilePicUrl ?? "",
-  };
-
   return (
     <Sidebar
       collapsible="icon"
@@ -138,7 +33,6 @@ export function AppSidebar({
       }}
       {...props}
     >
-      {/* Logo */}
       <SidebarHeader>
         <SidebarMenu>
           <SidebarMenuItem>
@@ -173,57 +67,13 @@ export function AppSidebar({
         </SidebarMenu>
       </SidebarHeader>
 
-      {/* Nav groups */}
-      <SidebarContent>
-        {groups.map((group, groupIndex) => (
-          <React.Fragment key={group.key}>
-            {groupIndex > 0 && (
-              <div className="px-3 py-1">
-                <Separator style={{ background: "var(--fab-border-md)" }} />
-              </div>
-            )}
-            <SidebarGroup className="py-0">
-              <SidebarGroupContent className="px-1.5 md:px-0">
-                <SidebarMenu>
-                  {group.items.map((item) => {
-                    const isActive =
-                      item.url !== "#" && pathname.startsWith(item.url);
+      <ProfileProvider preloadedProfile={preloadedProfile}>
+        <SidebarNavigation />
 
-                    return (
-                      <SidebarMenuItem key={item.title}>
-                        <SidebarMenuButton
-                          asChild
-                          tooltip={{ children: item.title, hidden: false }}
-                          isActive={isActive}
-                          className="px-2.5 md:px-2"
-                        >
-                          <Link
-                            href={item.url}
-                            style={
-                              {
-                                "--sidebar-icon-bg": item.iconBackground,
-                                "--sidebar-icon-color": item.iconColor,
-                              } as React.CSSProperties
-                            }
-                          >
-                            {item.icon}
-                            <span>{item.title}</span>
-                          </Link>
-                        </SidebarMenuButton>
-                      </SidebarMenuItem>
-                    );
-                  })}
-                </SidebarMenu>
-              </SidebarGroupContent>
-            </SidebarGroup>
-          </React.Fragment>
-        ))}
-      </SidebarContent>
-
-      {/* User */}
-      <SidebarFooter>
-        <NavUser user={user} />
-      </SidebarFooter>
+        <SidebarFooter>
+          <SidebarUserFooter />
+        </SidebarFooter>
+      </ProfileProvider>
     </Sidebar>
   );
 }
